Split up test_huggingface_clients
